Change kcm_audio from using generic get/set audio property functions to specific getter/setter functions for each object property that makes sense. Delete ALLEGRO_AUDIO_PROPERTY.
Change return types on many kcm_audio functions to return true on success
instead of zero on success.

+++ addons/kcm_audio/allegro5/internal/aintern_kcm_audio.h
@@ -168,7 +168,7 @@ struct ALLEGRO_STREAM {
    ALLEGRO_SAMPLE_INSTANCE       spl;
                         /* ALLEGRO_STREAM is derived from ALLEGRO_SAMPLE_INSTANCE. */
 
-   size_t               buf_count;
+   unsigned int         buf_count;
                         /* The stream buffer is divided into a number of
                          * fragments; this is the number of fragments.
                          */
